A start at a multi-version release pattern.
This commit moves everything that was in the kitchen/ source dir into a
new kitchen2/kitchen/ source dir.  Eventually, we'll merge the py3port
branch into a kitchen3/kitchen/ source dir parallel to it.

I added both source folders to ``MANIFEST.in`` so when we create a
tarball with ``python setup.py sdist``, **both** source directories will
be included -- kitchen2 and kitchen3.

When the user downloads and installs the package with ``pip install
kitchen``, it will download the tarball with both sources, unpack it,
and run the single setup.py.  setup.py has new logic in it now that will
detect if the user is running python2 - if it is, then it will install
kitchen from the kitchen2/ directory.  If they're running python3 it will
install kitchen from the kitchen3/ directory.

Note that, the imports for users do not change at all.  For both python2
and python3, kitchen can still be accessed with

```python
>>> import kitchen
```

This commit doesn't actually include any of the python3 bits, but it
stubs out one approach we could use which I'm submitting for feedback
and potential usage.